Assignment #1: Remove files generated by the script from the repo to pass autograder
Because of the past failures of passing the autograder, I assumed it was necessary to commit the generated files, too. It's quite the opposite, actually. If dir2/file3.txt is already in git, that means that file will contain two lines of "Hello, World!", and the rule # 4 of the autograder will fail.